FarisTours/requirements.txt
asgiref==3.3.4
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
Django==3.2.2
django-admin-interface==0.16.3
django-colorfield==0.4.1
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-grappelli==2.15.1
django-quill==0.1.0
django-quill-editor==0.1.22
django-recaptcha==2.0.6
django-widget-tweaks==1.4.8
google-api-core==1.26.3
google-api-python-client==2.4.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
oauth2client==4.1.3
packaging==20.9
Pillow==8.2.0
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
stripe==2.56.0
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.4
